vl.c: Report unknown machines correctly
I was recently typing in a QEMU command line, and ended up with
something like

    qemu-system-riscv64 -machine virt ... -M 8G

which is, in retrospect, obviously incorrect: there is no "8G" machine.
I should have typed something like

    qemu-system-riscv64 -machine virt ... -m 8G

but since QEMU was giving me the excessively unhelpful error message

    qemu-system-riscv64: -machine virt: unsupported machine type
    Use -machine help to list supported machines

I had to spend a few minutes scratching my head to figure out what was
going on.  For some reason I felt like I'd done that before, so I
figured I'd take a whack at fixing the problem this time.  It turns out
the error reporting for non-existant machines is just wrong: the invalid
machine is detected after we've lost the argument pointer, so it just
prints out the first instance of "-machine" instead of the one we're
actually looking for.

I've fixed this by just printing out "-machine $NAME" directly, but I
feel like there's a better way to do this.  Specifically, my issue is
that it always prints out "-machine" instead of "-M" -- that's actually
a regression for users just passing a single invalid machine via "-M",
which I assume is the more common case.

I'm not sure how to do this right, though, and my flight is boarding so
I figured I'd send this out as a way to ask the question.  I didn't have
time to run the test suite or figure out how to add a test for this, as
I'm assuming there's a better way to do it.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>